moves logos to own files in own folder
