+++ hviktor-angular/projects/hviktor/src/logo/index.ts
@@ -1,3 +1,5 @@
 export { LOGOS } from './logo-paths';
-export type { LogoCompany, LogoDefinition, LogoPath, LogoSize } from './logo-paths';
+export type { LogoCompany } from './logo-paths';
+export { buildLogo } from './logo-shared';
+export type { LogoDefinition, LogoPath, LogoSize } from './logo-types';
 export { HviLogo } from './logo.component';
