+++ hviktor-angular/projects/hviktor/src/logo/logo-paths.ts
@@ -1,117 +1,8 @@
-/** A single SVG path with fill type */
-export interface LogoPath {
-  /** SVG path data string */
-  d: string;
-  /**
-   * 'accent' = always #6CACE4 (blue dots)
-   * 'themed' = switches between dark (#003087) and light (white) via currentColor
-   */
-  fill: 'themed' | 'accent';
-}
-
-/** Defines the SVG structure for a single logo */
-export interface LogoDefinition {
-  /** SVG viewBox attribute */
-  viewBox: string;
-  /** Default width */
-  width: number;
-  /** Default height */
-  height: number;
-  /** Accessible label for the logo */
-  label: string;
-  /** SVG path elements */
-  paths: LogoPath[];
-}
-
-/** Available logo sizes */
-export type LogoSize = 'sm' | 'md' | 'lg';
-
-// ---------------------------------------------------------------------------
-// Shared paths for "dots-first" layout (all logos except Helse Vest, SAV, HBE and SUS)
-// Layout: [dots] [HELSE] [company name]
-// ---------------------------------------------------------------------------
-
-/** Accent dots — leading position (dots-first layout) */
-const DOTS: LogoPath[] = [
-  // Top
-  {
-    d: 'M39.9945 9.52374C39.9945 13.6821 36.6286 17.0493 32.4743 17.0493C28.3201 17.0493 24.9594 13.6839 24.9594 9.52374C24.9594 5.36354 28.3216 2 32.4725 2C36.6234 2 39.9926 5.3672 39.9926 9.52374H39.9945Z',
-    fill: 'accent',
-  },
-  // Center
-  {
-    d: 'M39.9945 32.482C39.9945 36.6331 36.6286 40.0021 32.4743 40.0021C28.3201 40.0021 24.9594 36.6331 24.9594 32.482C24.9594 28.331 28.3216 24.9531 32.4743 24.9531C36.627 24.9531 39.9945 28.3221 39.9945 32.482Z',
-    fill: 'accent',
-  },
-  // Right
-  {
-    d: 'M62.9656 32.482C62.9656 36.6331 59.5978 40.0021 55.447 40.0021C51.2961 40.0021 47.9283 36.6331 47.9283 32.482C47.9283 28.331 51.2943 24.9531 55.447 24.9531C59.5997 24.9531 62.9656 28.3221 62.9656 32.482Z',
-    fill: 'accent',
-  },
-  // Bottom
-  {
-    d: 'M39.9945 55.4799C39.9945 59.6346 36.6286 63 32.4743 63C28.3201 63 24.9594 59.6328 24.9594 55.4799C24.9594 51.327 28.3216 47.9492 32.4743 47.9492C36.627 47.9492 39.9945 51.3234 39.9945 55.4799Z',
-    fill: 'accent',
-  },
-];
-
-/** Themed dot — dots-first layout */
-const THEMED_DOT: LogoPath[] = [
-  {
-    d: 'M17.0373 32.482C17.0373 36.6331 13.6732 40.0021 9.51863 40.0021C5.36409 40.0021 2 36.6331 2 32.482C2 28.331 5.36775 24.9531 9.51863 24.9531C13.6695 24.9531 17.0373 28.3221 17.0373 32.482Z',
-    fill: 'themed',
-  },
-];
-
-/** Text "HELSE" — dots-first layout */
-const HELSE: LogoPath[] = [
-  // H
-  {
-    d: 'M100.095 44.2324V33.3638H86.1084V44.2324H83.1603V20.4799H86.1084V30.6648H100.095V20.4799H103.043V44.2324H100.095Z',
-    fill: 'themed',
-  },
-  // E
-  {
-    d: 'M110.883 44.2324V20.4799H124.222V22.9633H113.831V30.8808H122.424V33.3641H113.831V41.7494H124.546V44.2327H110.883V44.2324Z',
-    fill: 'themed',
-  },
-  // L
-  {
-    d: 'M130.263 44.2324V20.4799H133.211V41.7494H143.818V44.2327H130.263V44.2324Z',
-    fill: 'themed',
-  },
-  // S
-  {
-    d: 'M160.035 24.5105C158.812 23.7907 158.309 23.5388 157.554 23.2869C156.655 22.999 155.216 22.783 154.282 22.783C151.585 22.783 149.751 24.1146 149.751 26.1301C149.751 27.8936 150.794 28.793 155.612 31.3483C157.23 32.2121 158.488 33.0039 159.351 33.7597C160.502 34.7314 161.185 36.3509 161.185 38.0785C161.185 42.145 158.165 44.7003 153.383 44.7003C151.765 44.7003 150.254 44.4124 148.996 43.9085C148.097 43.5486 147.486 43.1887 145.94 42.145L147.45 39.8057C148.78 40.7054 149.355 41.0293 150.182 41.3532C151.225 41.7491 152.376 41.9651 153.598 41.9651C156.403 41.9651 158.093 40.6334 158.093 38.3661C158.093 37.6823 157.985 37.1064 157.733 36.6745C157.338 35.9547 156.655 35.2709 155.72 34.7311C155.324 34.5151 153.994 33.8313 151.765 32.6077C148.205 30.7002 146.695 28.8287 146.695 26.3814C146.695 22.6028 149.824 20.0835 154.534 20.0835C155.9 20.0835 157.302 20.2634 158.524 20.6593C159.495 20.9832 160.071 21.2711 161.401 22.0629L160.035 24.5099V24.5105Z',
-    fill: 'themed',
-  },
-  // E
-  {
-    d: 'M167.262 44.2324V20.4799H180.602V22.9633H170.21V30.8808H178.804V33.3641H170.21V41.7494H180.925V44.2327H167.262V44.2324Z',
-    fill: 'themed',
-  },
-];
-
-/** Builds a dots-first logo: [dots] [themed dot] [HELSE] [unique company text] */
-function buildLogo(
-  label: string,
-  companyPaths: LogoPath[],
-  viewBox = '0 0 240 61',
-  width = 240,
-  height = 61,
-): LogoDefinition {
-  return {
-    viewBox,
-    width,
-    height,
-    label,
-    paths: [...DOTS, ...THEMED_DOT, ...HELSE, ...companyPaths],
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Logo registry
-// ---------------------------------------------------------------------------
+import { LogoDefinition } from './logo-types';
+import { LOGO_HBE } from './logos/logo-hbe';
+import { LOGO_HBE_HUS } from './logos/logo-hbe-hus';
+import { LOGO_HVE } from './logos/logo-hve';
+import { LOGO_HVIKT } from './logos/logo-hvikt';
 
 /** Available company identifiers for the logo component */
 export type LogoCompany = keyof typeof LOGOS;
@@ -119,152 +10,12 @@ export type LogoCompany = keyof typeof LOGOS;
 /**
  * Path data and metadata for all logos.
  *
- * Each logo is defined as a standalone object with its own paths because
- * the SVG layouts differ — "Helse Vest" has dots in the middle while all
- * other logos have dots as leading elements.
+ * Individual logo definitions live in their own files (logo-hve.ts, logo-hvikt.ts, etc.)
+ * to keep file sizes manageable. This registry assembles them.
  */
 export const LOGOS = {
-  hvikt: buildLogo(
-    'Helse Vest IKT logo',
-    [
-      // V
-      {
-        d: 'M203.757 44.9159H200.593L191.784 20.4796H196.746L202.175 36.5306C202.283 36.8905 202.391 37.2864 202.427 37.6463C202.427 37.3224 202.463 37.0704 202.643 36.5306L207.821 20.4796H212.567L203.757 44.9159Z',
-        fill: 'themed',
-      },
-      // E
-      {
-        d: 'M215.911 44.2324V20.4799H230.365V24.3306H220.442V30.1966H228.568V34.0473H220.442V40.3815H230.833V44.2321H215.911L215.911 44.2324Z',
-        fill: 'themed',
-      },
-      // S
-      {
-        d: 'M248.415 26.1301C245.898 25.0144 243.848 24.4386 242.266 24.4386C240.145 24.4386 238.779 25.3023 238.779 26.6339C238.779 27.6417 239.354 28.2535 241.044 29.1532C244.675 31.0967 247.049 32.6083 248.163 33.7237C249.17 34.7314 249.781 36.315 249.781 37.9705C249.781 42.217 246.689 44.7003 241.44 44.7003C240.72 44.7003 240.001 44.6643 239.318 44.5564C237.629 44.3404 236.837 44.0525 234.356 42.9008L232.81 42.181L234.716 38.0062C236.442 38.9419 237.161 39.3018 238.096 39.6617C239.282 40.0936 240.577 40.3455 241.512 40.3455C243.597 40.3455 244.963 39.4098 244.963 37.9702C244.963 36.9625 244.424 36.2787 242.878 35.4149L239.21 33.3275C235.003 30.9161 233.709 29.3326 233.709 26.7416C233.709 25.5539 234.068 24.3303 234.644 23.4308C235.039 22.819 235.651 22.2072 236.406 21.6673C238.132 20.4797 239.606 20.0838 242.374 20.0838C243.956 20.0838 245.359 20.2637 246.833 20.6956C247.983 21.0555 248.631 21.3074 250.212 22.0632L248.415 26.1301Z',
-        fill: 'themed',
-      },
-      // T
-      {
-        d: 'M263.194 24.3306V44.2324H258.663V24.3306H251.544V20.4799H270.313V24.3306H263.194Z',
-        fill: 'themed',
-      },
-      // I
-      { d: 'M283.545 44.2324V20.4799H286.493V44.2324H283.545Z', fill: 'themed' },
-      // K
-      {
-        d: 'M294.331 44.2324V20.4799H297.28V44.2324H294.331ZM306.988 44.2324L297.496 32.2481L306.772 20.4799H310.152L300.875 32.0324L310.619 44.2324H306.988Z',
-        fill: 'themed',
-      },
-      // T
-      {
-        d: 'M322.089 22.963V44.2324H319.141V22.963H311.231V20.4796H330V22.963H322.089Z',
-        fill: 'themed',
-      },
-    ],
-    '0 0 336 65',
-    336,
-    65,
-  ),
-  hve: {
-    viewBox: '0 0 240 61',
-    width: 240,
-    height: 61,
-    label: 'Helse Vest logo',
-    paths: [
-      // Accent dots (center)
-      {
-        d: 'M136.787 7.41402C136.787 11.5106 133.46 14.8302 129.357 14.8302C125.254 14.8302 121.932 11.5106 121.932 7.41402C121.932 3.31748 125.259 0 129.357 0C133.456 0 136.787 3.31964 136.787 7.41402Z',
-        fill: 'accent',
-      },
-      {
-        d: 'M136.787 30.0374C136.787 34.1318 133.46 37.4493 129.357 37.4493C125.254 37.4493 121.932 34.1297 121.932 30.0374C121.932 25.9452 125.259 22.6188 129.357 22.6188C133.456 22.6188 136.787 25.9409 136.787 30.0374Z',
-        fill: 'accent',
-      },
-      {
-        d: 'M159.48 30.0374C159.48 34.1318 156.153 37.4493 152.053 37.4493C147.952 37.4493 144.625 34.1297 144.625 30.0374C144.625 25.9452 147.95 22.6188 152.053 22.6188C156.156 22.6188 159.48 25.9409 159.48 30.0374Z',
-        fill: 'accent',
-      },
-      {
-        d: 'M136.787 52.702C136.787 56.7942 133.46 60.116 129.357 60.116C125.254 60.116 121.932 56.7942 121.932 52.702C121.932 48.6098 125.259 45.2833 129.357 45.2833C133.456 45.2833 136.787 48.6033 136.787 52.702Z',
-        fill: 'accent',
-      },
-      // Themed dot
-      {
-        d: 'M114.11 30.0374C114.11 34.1318 110.785 37.4493 106.682 37.4493C102.579 37.4493 99.2563 34.1297 99.2563 30.0374C99.2563 25.9452 102.581 22.6188 106.682 22.6188C110.783 22.6188 114.11 25.9409 114.11 30.0374Z',
-        fill: 'themed',
-      },
-      // HELSE
-      {
-        d: 'M14.3338 40.0273V30.8515H2.49643V40.0273H0V19.9711H2.49643V28.5712H14.3338V19.9711H16.8302V40.0273H14.3338Z',
-        fill: 'themed',
-      },
-      {
-        d: 'M23.8173 40.0273V19.9711H35.1068V22.0675H26.3115V28.7529H33.5835V30.8493H26.3115V37.9287H35.3809V40.0251H23.8173V40.0273Z',
-        fill: 'themed',
-      },
-      { d: 'M40.5774 40.0273V19.9711H43.0738V37.9309H52.0511V40.0273H40.5774Z', fill: 'themed' },
-      {
-        d: 'M66.1307 23.374C65.0962 22.768 64.6688 22.5535 64.0311 22.3411C63.2707 22.0982 62.0542 21.9165 61.2606 21.9165C58.979 21.9165 57.4253 23.0413 57.4253 24.744C57.4253 26.2322 58.3085 26.9914 62.3872 29.1491C63.757 29.8777 64.8222 30.5475 65.5523 31.1864C66.5253 32.0069 67.1038 33.3747 67.1038 34.8344C67.1038 38.2658 64.5482 40.4257 60.5002 40.4257C59.1303 40.4257 57.8525 40.1827 56.7873 39.7562C56.0269 39.4519 55.5095 39.1498 54.2013 38.2658L55.4788 36.292C56.6053 37.0534 57.092 37.327 57.7933 37.6003C58.6746 37.9331 59.6499 38.1148 60.6843 38.1148C63.058 38.1148 64.4893 36.99 64.4893 35.0752C64.4893 34.4974 64.3972 34.0139 64.1845 33.6483C63.8513 33.0423 63.2707 32.4644 62.4814 32.0069C62.1463 31.8256 61.0197 31.2477 59.1346 30.2148C56.1211 28.604 54.8433 27.0242 54.8433 24.9585C54.8433 21.7679 57.491 19.6405 61.4776 19.6405C62.6348 19.6405 63.8206 19.7937 64.8572 20.1265C65.6794 20.4001 66.1657 20.6431 67.2923 21.3126L66.1351 23.3783L66.1307 23.374Z',
-        fill: 'themed',
-      },
-      {
-        d: 'M72.6051 40.0273V19.9711H83.8947V22.0675H75.0994V28.7529H82.3713V30.8493H75.0994V37.9287H84.1687V40.0251H72.6051V40.0273Z',
-        fill: 'themed',
-      },
-      // VEST
-      {
-        d: 'M182.5 40.6114H179.819L172.365 19.9773H176.565L181.161 33.5298C181.253 33.8341 181.343 34.169 181.373 34.4729C181.373 34.1996 181.406 33.9851 181.558 33.5298L185.939 19.9773H189.956L182.502 40.6114H182.5Z',
-        fill: 'themed',
-      },
-      {
-        d: 'M193.079 40.0338V19.9776H205.314V23.2295H196.915V28.1818H203.792V31.4337H196.915V36.7841H205.71V40.0338H193.079Z',
-        fill: 'themed',
-      },
-      {
-        d: 'M220.886 24.7505C218.753 23.8073 217.019 23.3214 215.681 23.3214C213.885 23.3214 212.728 24.0503 212.728 25.1751C212.728 26.0263 213.217 26.5426 214.646 27.3022C217.719 28.9433 219.726 30.2191 220.671 31.1601C221.523 32.0113 222.041 33.3484 222.041 34.7468C222.041 38.3336 219.424 40.43 214.979 40.43C214.372 40.43 213.763 40.3994 213.184 40.3074C211.755 40.1257 211.085 39.8831 208.985 38.909L207.676 38.303L209.287 34.7775C210.749 35.5674 211.356 35.8716 212.15 36.1759C213.153 36.5415 214.249 36.7535 215.04 36.7535C216.805 36.7535 217.962 35.9657 217.962 34.749C217.962 33.8978 217.504 33.3221 216.196 32.5913L213.092 30.8276C209.531 28.7922 208.435 27.4554 208.435 25.2671C208.435 24.2645 208.739 23.2316 209.226 22.4724C209.559 21.9561 210.079 21.4395 210.718 20.9842C212.178 19.9819 213.425 19.6473 215.77 19.6473C217.109 19.6473 218.295 19.8006 219.544 20.1636C220.518 20.4679 221.066 20.6802 222.407 21.3169L220.886 24.7505Z',
-        fill: 'themed',
-      },
-      {
-        d: 'M233.69 23.2295V40.0338H229.854V23.2295H223.829V19.9776H239.715V23.2295H233.69Z',
-        fill: 'themed',
-      },
-    ],
-  },
-  hbe: buildLogo(
-    'Helse Bergen logo',
-    [
-      // B
-      {
-        d: 'M207.347 43.9684H197.525V19.8636H206.251C209.391 19.8636 210.962 20.2654 212.276 21.4707C213.444 22.4934 214.175 24.21 214.175 25.7802C214.175 27.1316 213.663 28.556 212.787 29.5784C211.728 30.7836 210.743 31.2219 208.844 31.3315C212.495 31.4046 214.795 33.7786 214.795 37.5402C214.795 41.704 212.166 43.9684 207.347 43.9684ZM205.339 23.5157H202.345V29.7248H205.083C207.676 29.7248 209.063 28.6291 209.063 26.584C209.063 24.6483 207.676 23.516 205.339 23.516V23.5157ZM205.339 33.6325H202.345V40.2798H205.339C206.763 40.2798 207.42 40.1703 208.113 39.8416C209.136 39.3302 209.756 38.271 209.756 37.0292C209.756 34.8378 208.186 33.6325 205.339 33.6325Z',
-        fill: 'themed',
-      },
-      // E
-      {
-        d: 'M219.323 43.9684V19.8636H234.001V23.7714H223.924V29.7245H232.175V33.6322H223.924V40.0604H234.476V43.9681H219.323L219.323 43.9684Z',
-        fill: 'themed',
-      },
-      // R
-      {
-        d: 'M251.417 43.9684L245.904 34.3265H243.604V43.9684H239.003V19.8636H247.036C250.541 19.8636 252.148 20.3019 253.572 21.6533C254.886 22.8586 255.543 24.6117 255.543 26.8762C255.543 28.7389 255.105 30.2361 254.229 31.4049C253.316 32.5736 252.476 33.085 250.468 33.6328L256.456 43.9687H251.417V43.9684ZM246.598 23.7714H243.604V30.4187H246.598C249.519 30.4187 250.797 29.3961 250.797 26.9855C250.797 24.6845 249.592 23.7714 246.598 23.7714Z',
-        fill: 'themed',
-      },
-      // G
-      {
-        d: 'M277.78 43.6397C275.224 44.2241 273.727 44.4433 271.719 44.4433C263.723 44.4433 258.173 39.3299 258.173 31.8794C258.173 28.227 259.56 25.0497 262.153 22.6757C264.635 20.4115 267.41 19.4619 271.573 19.4619C273.325 19.4619 275.078 19.6445 276.611 20.0098C277.78 20.302 278.51 20.5576 280.189 21.2878L279.094 25.1228C278.218 24.7941 277.56 24.5749 277.122 24.4288C275.406 23.8445 273.69 23.5523 272.157 23.5523C266.753 23.5523 262.992 26.8759 262.992 31.6602C262.992 36.5907 267.118 40.3526 272.485 40.3526C273.581 40.3526 274.201 40.243 275.625 39.8413V33.1939H280.226V43.0551L277.78 43.6394V43.6397Z',
-        fill: 'themed',
-      },
-      // E
-      {
-        d: 'M286.178 43.9684V19.8636H300.856V23.7714H290.779V29.7245H299.03V33.6322H290.779V40.0604H301.331V43.9681H286.178L286.178 43.9684Z',
-        fill: 'themed',
-      },
-      // N
-      {
-        d: 'M321.923 44.1876L311.59 28.5925L310.787 27.1316C310.641 26.8394 310.568 26.6933 310.458 26.2915V43.9684H306.077V19.8636H311.371L321.119 35.13C321.813 36.2257 322.105 36.9927 322.397 37.7597V19.8636H326.743V44.1876H321.923Z',
-        fill: 'themed',
-      },
-    ],
-    '0 0 330 65',
-    330,
-    65,
-  ),
+  hvikt: LOGO_HVIKT,
+  hve: LOGO_HVE,
+  hbe: LOGO_HBE,
+  'hbe-hus': LOGO_HBE_HUS,
 } satisfies Record<string, LogoDefinition>;
