+++ hviktor-angular/projects/hviktor/src/logo/logo-shared.ts
@@ -1,0 +1,84 @@
+import { LogoDefinition, LogoPath } from './logo-types';
+
+// ---------------------------------------------------------------------------
+// Shared paths for "dots-first" layout (all logos except Helse Vest, SAV, HBE and SUS)
+// Layout: [dots] [HELSE] [company name]
+// ---------------------------------------------------------------------------
+
+/** Accent dots — leading position (dots-first layout) */
+const DOTS: LogoPath[] = [
+  // Top
+  {
+    d: 'M39.9945 9.52374C39.9945 13.6821 36.6286 17.0493 32.4743 17.0493C28.3201 17.0493 24.9594 13.6839 24.9594 9.52374C24.9594 5.36354 28.3216 2 32.4725 2C36.6234 2 39.9926 5.3672 39.9926 9.52374H39.9945Z',
+    fill: 'accent',
+  },
+  // Center
+  {
+    d: 'M39.9945 32.482C39.9945 36.6331 36.6286 40.0021 32.4743 40.0021C28.3201 40.0021 24.9594 36.6331 24.9594 32.482C24.9594 28.331 28.3216 24.9531 32.4743 24.9531C36.627 24.9531 39.9945 28.3221 39.9945 32.482Z',
+    fill: 'accent',
+  },
+  // Right
+  {
+    d: 'M62.9656 32.482C62.9656 36.6331 59.5978 40.0021 55.447 40.0021C51.2961 40.0021 47.9283 36.6331 47.9283 32.482C47.9283 28.331 51.2943 24.9531 55.447 24.9531C59.5997 24.9531 62.9656 28.3221 62.9656 32.482Z',
+    fill: 'accent',
+  },
+  // Bottom
+  {
+    d: 'M39.9945 55.4799C39.9945 59.6346 36.6286 63 32.4743 63C28.3201 63 24.9594 59.6328 24.9594 55.4799C24.9594 51.327 28.3216 47.9492 32.4743 47.9492C36.627 47.9492 39.9945 51.3234 39.9945 55.4799Z',
+    fill: 'accent',
+  },
+];
+
+/** Themed dot — dots-first layout */
+const THEMED_DOT: LogoPath[] = [
+  {
+    d: 'M17.0373 32.482C17.0373 36.6331 13.6732 40.0021 9.51863 40.0021C5.36409 40.0021 2 36.6331 2 32.482C2 28.331 5.36775 24.9531 9.51863 24.9531C13.6695 24.9531 17.0373 28.3221 17.0373 32.482Z',
+    fill: 'themed',
+  },
+];
+
+/** Text "HELSE" — dots-first layout */
+const HELSE: LogoPath[] = [
+  // H
+  {
+    d: 'M100.095 44.2324V33.3638H86.1084V44.2324H83.1603V20.4799H86.1084V30.6648H100.095V20.4799H103.043V44.2324H100.095Z',
+    fill: 'themed',
+  },
+  // E
+  {
+    d: 'M110.883 44.2324V20.4799H124.222V22.9633H113.831V30.8808H122.424V33.3641H113.831V41.7494H124.546V44.2327H110.883V44.2324Z',
+    fill: 'themed',
+  },
+  // L
+  {
+    d: 'M130.263 44.2324V20.4799H133.211V41.7494H143.818V44.2327H130.263V44.2324Z',
+    fill: 'themed',
+  },
+  // S
+  {
+    d: 'M160.035 24.5105C158.812 23.7907 158.309 23.5388 157.554 23.2869C156.655 22.999 155.216 22.783 154.282 22.783C151.585 22.783 149.751 24.1146 149.751 26.1301C149.751 27.8936 150.794 28.793 155.612 31.3483C157.23 32.2121 158.488 33.0039 159.351 33.7597C160.502 34.7314 161.185 36.3509 161.185 38.0785C161.185 42.145 158.165 44.7003 153.383 44.7003C151.765 44.7003 150.254 44.4124 148.996 43.9085C148.097 43.5486 147.486 43.1887 145.94 42.145L147.45 39.8057C148.78 40.7054 149.355 41.0293 150.182 41.3532C151.225 41.7491 152.376 41.9651 153.598 41.9651C156.403 41.9651 158.093 40.6334 158.093 38.3661C158.093 37.6823 157.985 37.1064 157.733 36.6745C157.338 35.9547 156.655 35.2709 155.72 34.7311C155.324 34.5151 153.994 33.8313 151.765 32.6077C148.205 30.7002 146.695 28.8287 146.695 26.3814C146.695 22.6028 149.824 20.0835 154.534 20.0835C155.9 20.0835 157.302 20.2634 158.524 20.6593C159.495 20.9832 160.071 21.2711 161.401 22.0629L160.035 24.5099V24.5105Z',
+    fill: 'themed',
+  },
+  // E
+  {
+    d: 'M167.262 44.2324V20.4799H180.602V22.9633H170.21V30.8808H178.804V33.3641H170.21V41.7494H180.925V44.2327H167.262V44.2324Z',
+    fill: 'themed',
+  },
+];
+
+/** Builds a dots-first logo: [dots] [themed dot] [HELSE] [unique company text] */
+export function buildLogo(
+  label: string,
+  companyPaths: LogoPath[],
+  viewBox = '0 0 240 61',
+  width = 240,
+  height = 61,
+): LogoDefinition {
+  return {
+    viewBox,
+    width,
+    height,
+    label,
+    paths: [...DOTS, ...THEMED_DOT, ...HELSE, ...companyPaths],
+  };
+}
