+++ hviktor-angular/projects/hviktor/src/logo/logos/logo-hbe-hus.ts
@@ -1,0 +1,189 @@
+import { LogoDefinition } from '../logo-types';
+
+export const LOGO_HBE_HUS: LogoDefinition = {
+  viewBox: '0 0 348 65',
+  width: 348,
+  height: 65,
+  label: 'Haukeland universitetssjukehus logo',
+  paths: [
+    // Accent dots
+    {
+      d: 'M38.6034 9.41943C38.6034 13.3194 35.4433 16.4766 31.5395 16.4766C27.6358 16.4766 24.4756 13.3194 24.4756 9.41943C24.4756 5.51943 27.6358 2.36229 31.5395 2.36229C35.4433 2.36229 38.6034 5.51943 38.6034 9.41943Z',
+      fill: 'accent',
+    },
+    {
+      d: 'M38.6034 30.9623C38.6034 34.8623 35.4433 38.0194 31.5395 38.0194C27.6358 38.0194 24.4756 34.8623 24.4756 30.9623C24.4756 27.0623 27.6358 23.9052 31.5395 23.9052C35.4433 23.9052 38.6034 27.0623 38.6034 30.9623Z',
+      fill: 'accent',
+    },
+    {
+      d: 'M60.198 30.9623C60.198 34.8623 57.0379 38.0194 53.1341 38.0194C49.2304 38.0194 46.0702 34.8623 46.0702 30.9623C46.0702 27.0623 49.2304 23.9052 53.1341 23.9052C57.0379 23.9052 60.198 27.0623 60.198 30.9623Z',
+      fill: 'accent',
+    },
+    {
+      d: 'M38.6034 52.5361C38.6034 56.4361 35.4433 59.5932 31.5395 59.5932C27.6358 59.5932 24.4756 56.4361 24.4756 52.5361C24.4756 48.6361 27.6358 45.479 31.5395 45.479C35.4433 45.479 38.6034 48.6361 38.6034 52.5361Z',
+      fill: 'accent',
+    },
+    // Themed dot
+    {
+      d: 'M17.0088 30.9623C17.0088 34.8623 13.8487 38.0194 9.94491 38.0194C6.04116 38.0194 2.88098 34.8623 2.88098 30.9623C2.88098 27.0623 6.04116 23.9052 9.94491 23.9052C13.8487 23.9052 17.0088 27.0623 17.0088 30.9623Z',
+      fill: 'themed',
+    },
+    // HELSE
+    {
+      d: 'M95.2078 38.0504V27.867H82.1024V38.0504H79.345V15.8266H82.1024V25.3599H95.2078V15.8266H97.9652V38.0504H95.2078Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M105.308 38.0504V15.8266H117.825V18.148H108.096V25.5456H116.152V27.867H108.096V35.7289H118.135V38.0504H105.308Z',
+      fill: 'themed',
+    },
+    { d: 'M123.464 38.0504V15.8266H126.221V35.7289H136.166V38.0504H123.464Z', fill: 'themed' },
+    {
+      d: 'M151.378 19.5718C150.232 18.8909 149.767 18.6742 149.055 18.4266C148.218 18.148 146.855 17.9623 145.988 17.9623C143.447 17.9623 141.743 19.2004 141.743 21.0885C141.743 22.729 142.734 23.5647 147.227 25.9789C148.745 26.7837 149.922 27.5266 150.728 28.2385C151.812 29.1361 152.432 30.6528 152.432 32.2932C152.432 36.1004 149.612 38.4837 145.12 38.4837C143.602 38.4837 142.177 38.2051 140.999 37.7409C140.163 37.4004 139.574 37.0599 138.118 36.1004L139.543 33.9028C140.782 34.7385 141.34 35.048 142.115 35.3575C143.106 35.729 144.16 35.9147 145.306 35.9147C147.939 35.9147 149.519 34.6766 149.519 32.5409C149.519 31.8909 149.427 31.3647 149.179 30.9623C148.807 30.2813 148.156 29.6623 147.289 29.1361C146.917 28.9194 145.678 28.3004 143.571 27.1551C140.225 25.3599 138.831 23.6266 138.831 21.3361C138.831 17.8075 141.774 15.4551 146.173 15.4551C147.444 15.4551 148.776 15.6099 149.922 15.9813C150.821 16.2909 151.378 16.5385 152.618 17.2813L151.378 19.5718Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M158.133 38.0504V15.8266H170.649V18.148H160.89V25.5456H168.945V27.867H160.89V35.7289H170.928V38.0504H158.133Z',
+      fill: 'themed',
+    },
+    // BERGEN
+    {
+      d: 'M192.956 38.0504H183.879V15.8266H191.934C194.846 15.8266 196.272 16.198 197.511 17.3123C198.595 18.2409 199.277 19.8504 199.277 21.2742C199.277 22.5123 198.812 23.8432 198.007 24.7718C197.015 25.8861 196.117 26.2885 194.382 26.3813C197.759 26.4432 199.865 28.6409 199.865 32.1075C199.834 35.9456 197.418 38.0504 192.956 38.0504ZM191.128 19.1694H188.371V24.8956H190.912C193.297 24.8956 194.598 23.8742 194.598 21.9861C194.537 20.2218 193.266 19.1694 191.128 19.1694ZM191.128 28.517H188.371V34.6456H191.128C192.43 34.6456 193.049 34.5528 193.7 34.2432C194.629 33.7789 195.218 32.7885 195.218 31.6432C195.187 29.6313 193.731 28.517 191.128 28.517Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M204.017 38.0504V15.8266H217.556V19.417H208.262V24.8956H215.883V28.4861H208.262V34.398H217.99V37.9885L204.017 38.0504Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M233.636 38.0504L228.555 29.167H226.417V38.0504H222.173V15.8266H229.577C232.799 15.8266 234.287 16.2289 235.619 17.467C236.827 18.5813 237.447 20.1909 237.447 22.2956C237.447 23.998 237.044 25.3909 236.239 26.4742C235.402 27.5575 234.627 28.0218 232.769 28.517L238.283 38.0504H233.636ZM229.175 19.417H226.417V25.5456H229.175C231.87 25.5456 233.047 24.5861 233.047 22.3885C233.078 20.2528 231.963 19.417 229.175 19.417Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M257.957 37.7409C255.602 38.267 254.208 38.4837 252.349 38.4837C244.976 38.4837 239.832 33.7789 239.832 26.9075C239.832 23.5337 241.103 20.6242 243.519 18.4266C245.812 16.3528 248.384 15.4551 252.225 15.4551C253.836 15.4551 255.447 15.6099 256.873 15.9504C257.957 16.2289 258.639 16.4456 260.188 17.1266L259.165 20.6551C258.36 20.3456 257.74 20.1599 257.337 20.0051C255.757 19.479 254.177 19.2004 252.752 19.2004C247.764 19.2004 244.294 22.2647 244.294 26.6909C244.294 31.2409 248.105 34.7075 253.062 34.7075C254.084 34.7075 254.642 34.6147 255.974 34.2432V28.1147H260.219V37.2147L257.957 37.7409Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M265.703 38.0504V15.8266H279.242V19.417H269.947V24.8956H277.569V28.4861H269.947V34.398H279.675V37.9885L265.703 38.0504Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M298.698 38.2361L289.156 23.8742L288.412 22.5123C288.288 22.2337 288.196 22.1099 288.103 21.7385V38.0504H284.044V15.8266H288.939L297.924 29.9099C298.575 30.9313 298.822 31.6123 299.101 32.3242V15.8266H303.098V38.267H298.698V38.2361Z',
+      fill: 'themed',
+    },
+    // Haukeland universitetssjukehus (second line)
+    {
+      d: 'M89.0734 59.098V53.0004H81.2039V59.098H79.5309V45.7266H81.2039V51.4528H89.0734V45.7266H90.7464V59.098H89.0734Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M101.187 59.3456C99.9171 59.2528 99.4834 58.9432 99.3904 58.0766C98.9257 58.479 98.7088 58.6337 98.43 58.8194C97.8723 59.129 97.1287 59.3456 96.4781 59.3456C94.991 59.3456 94.0305 58.3861 94.0305 56.9313C94.0305 55.6004 94.7431 54.7028 96.0753 54.3932L99.2665 53.6504V51.979C99.2665 51.2361 98.7088 50.8337 97.5315 50.8337C96.5711 50.8337 96.1373 51.0813 96.1373 51.6385C96.1373 51.7313 96.1683 51.8242 96.2303 51.979L94.6811 52.2885C94.5262 52.0409 94.5262 51.948 94.5262 51.7313C94.5262 50.5242 95.8585 49.6885 97.8413 49.6885C99.7003 49.6885 100.847 50.5551 100.847 51.979V57.6123C100.878 57.8599 101.125 58.0147 101.59 58.1075L101.187 59.3456ZM99.2665 54.8885L95.8275 55.7861C95.6106 56.2194 95.5796 56.4051 95.5796 56.7766C95.5796 57.5194 96.0753 57.9837 96.8189 57.9837C97.5005 57.9837 97.9652 57.767 99.2975 56.8075V54.8885H99.2665Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M110.327 59.098V57.2099C109.243 58.6956 108.065 59.4075 106.671 59.4075C105.06 59.4075 104.224 58.5099 104.224 56.8385V49.967H105.742V56.467C105.742 57.6123 106.052 58.0766 106.826 58.0766C107.787 58.0766 109.243 57.0551 110.296 55.6004V49.998H111.845V59.1599C111.876 59.098 110.327 59.098 110.327 59.098Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M115.191 59.098V44.3337H116.709V59.098H115.191ZM120.551 59.098L116.771 54.1147L120.551 49.9361H122.534L118.63 54.1456L122.565 59.098H120.551Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M125.013 54.3623C125.013 56.5908 126.159 57.9528 128.018 57.9528C128.823 57.9528 129.412 57.7361 130.403 57.0551L131.023 58.2623C129.66 59.098 128.885 59.3766 127.708 59.3766C125.168 59.3766 123.432 57.4266 123.432 54.548C123.432 51.6385 125.167 49.7194 127.77 49.7194C129.908 49.7194 131.302 51.2361 131.302 53.5266V54.4242L125.013 54.3623ZM127.646 51.0504C126.376 51.0504 125.477 51.8242 125.137 53.248H129.629C129.598 51.8242 128.885 51.0504 127.646 51.0504Z',
+      fill: 'themed',
+    },
+    { d: 'M133.688 59.098V44.3337H135.206V59.098H133.688Z', fill: 'themed' },
+    {
+      d: 'M145.12 59.3456C143.85 59.2528 143.416 58.9432 143.323 58.0766C142.858 58.479 142.641 58.6337 142.363 58.8194C141.805 59.129 141.061 59.3456 140.411 59.3456C138.924 59.3456 137.963 58.3861 137.963 56.9313C137.963 55.6004 138.676 54.7028 140.008 54.3932L143.199 53.6504V51.979C143.199 51.2361 142.641 50.8337 141.464 50.8337C140.504 50.8337 140.07 51.0813 140.07 51.6385C140.07 51.7313 140.101 51.8242 140.163 51.979L138.614 52.2885C138.459 52.0409 138.459 51.948 138.459 51.7313C138.459 50.5242 139.791 49.6885 141.774 49.6885C143.633 49.6885 144.779 50.5551 144.779 51.979V57.6123C144.81 57.8599 145.058 58.0147 145.523 58.1075L145.12 59.3456ZM143.199 54.8885L139.76 55.7861C139.543 56.2194 139.512 56.4051 139.512 56.7766C139.512 57.5194 140.008 57.9837 140.752 57.9837C141.433 57.9837 141.898 57.767 143.23 56.8075L143.199 54.8885Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M154.384 59.098V52.598C154.384 51.979 154.353 51.7932 154.229 51.5147C154.074 51.2052 153.733 51.0194 153.299 51.0194C152.37 51.0194 151.378 51.7004 149.767 53.4647V59.0671H148.249V49.9051H149.767V51.8242C150.387 51.1123 150.573 50.9266 151.038 50.5551C151.719 49.9671 152.649 49.6575 153.485 49.6575C155.096 49.6575 155.902 50.4932 155.902 52.1956V59.0671L154.384 59.098Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M165.289 59.098V57.9528C164.484 58.8504 163.399 59.3766 162.253 59.3766C159.991 59.3766 158.411 57.4885 158.411 54.7956C158.411 51.8551 160.363 49.7194 163.09 49.7194C164.05 49.7194 164.701 49.9051 165.289 50.3385V44.3647H166.839V59.129L165.289 59.098ZM165.289 53.1242C165.289 52.4123 165.258 52.2266 165.104 51.917C164.856 51.4218 164.205 51.1432 163.245 51.1432C161.293 51.1432 160.053 52.5361 160.053 54.7337C160.053 56.5909 161.076 57.9218 162.501 57.9218C163.059 57.9218 163.709 57.7051 164.298 57.3028C164.639 57.0551 164.856 56.8385 165.289 56.3123C165.289 56.2813 165.289 53.1242 165.289 53.1242Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M181.276 59.098V57.2099C180.192 58.6956 179.015 59.4075 177.62 59.4075C176.009 59.4075 175.173 58.5099 175.173 56.8385V49.967H176.691V56.467C176.691 57.6123 177.001 58.0766 177.775 58.0766C178.736 58.0766 180.192 57.0551 181.245 55.6004V49.998H182.794V59.1599H181.276V59.098Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M192.337 59.098V52.598C192.337 51.979 192.306 51.7932 192.182 51.5147C192.027 51.2052 191.686 51.0194 191.252 51.0194C190.323 51.0194 189.332 51.7004 187.721 53.4647V59.0671H186.202V49.9051H187.721V51.8242C188.34 51.1123 188.526 50.9266 188.991 50.5551C189.672 49.9671 190.602 49.6575 191.438 49.6575C193.049 49.6575 193.855 50.4932 193.855 52.1956V59.0671L192.337 59.098Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M198.038 47.7075C197.48 47.7075 197.046 47.2742 197.046 46.717C197.046 46.1599 197.48 45.7266 198.038 45.7266C198.595 45.7266 199.029 46.1599 199.029 46.717C199.029 47.2742 198.595 47.7075 198.038 47.7075ZM197.263 59.098V49.9361H198.781V59.098H197.263Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M205.721 59.3456H204.327L200.857 49.9051H202.53L205.07 57.3337L205.101 57.7051L205.132 57.3647L207.611 49.9051H209.284L205.721 59.3456Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M212.042 54.3623C212.042 56.5908 213.188 57.9528 215.047 57.9528C215.852 57.9528 216.441 57.7361 217.432 57.0551L218.052 58.2623C216.689 59.098 215.914 59.3766 214.737 59.3766C212.196 59.3766 210.461 57.4266 210.461 54.548C210.461 51.6385 212.196 49.7194 214.799 49.7194C216.937 49.7194 218.331 51.2361 218.331 53.5266V54.4242L212.042 54.3623ZM214.675 51.0504C213.405 51.0504 212.506 51.8242 212.165 53.248H216.658C216.658 51.8242 215.945 51.0504 214.675 51.0504Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M225.147 51.5766C224.837 51.3289 224.682 51.2051 224.496 51.2051C223.939 51.2051 223.195 52.0408 222.328 53.6194V59.098H220.809V49.9361H222.328V51.6075C223.474 50.1528 224.032 49.7194 224.713 49.7194C225.054 49.7194 225.488 49.9361 225.891 50.3385L225.147 51.5766Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M230.538 59.3456C230.166 59.3456 229.732 59.3147 229.268 59.2218C228.4 59.067 227.997 58.9123 227.099 58.448L227.718 57.179L228.183 57.4266C229.082 57.9218 229.949 58.0456 230.507 58.0456C231.436 58.0456 232.118 57.5504 232.118 56.8385C232.118 56.6218 232.025 56.3432 231.932 56.2194C231.56 55.6932 231.126 55.4147 229.515 54.7647C227.718 54.0218 227.13 53.4028 227.13 52.2885C227.13 50.679 228.462 49.6885 230.6 49.6885C230.972 49.6885 231.405 49.7194 231.777 49.7813C232.459 49.9051 232.799 50.029 233.574 50.4623L232.83 51.7004C232.242 51.329 231.994 51.2051 231.498 51.0813C231.126 50.9885 230.786 50.9266 230.445 50.9266C229.422 50.9266 228.741 51.3909 228.741 52.0718C228.741 52.2885 228.834 52.5051 228.958 52.629C229.329 53.0004 229.577 53.1242 231.219 53.8361C233.171 54.6409 233.698 55.2599 233.698 56.7147C233.729 58.3551 232.552 59.3456 230.538 59.3456Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M237.478 47.7075C236.92 47.7075 236.486 47.2742 236.486 46.717C236.486 46.1599 236.92 45.7266 237.478 45.7266C238.036 45.7266 238.469 46.1599 238.469 46.717C238.469 47.2742 238.005 47.7075 237.478 47.7075ZM236.703 59.098V49.9361H238.221V59.098H236.703Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M243.55 59.3456C242.497 59.3456 241.784 58.5099 241.784 57.3028V51.2051H240.235V50.3075L241.784 49.9361V47.2432H243.302V49.9051H245.719V51.1742H243.302V56.4051C243.302 56.498 243.333 57.2409 243.364 57.4266C243.395 57.6742 243.581 57.829 243.891 57.829C244.263 57.829 244.573 57.6742 245.223 57.179L245.843 58.2004C244.79 59.098 244.294 59.3456 243.55 59.3456Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M248.693 54.3623C248.693 56.5908 249.84 57.9528 251.699 57.9528C252.504 57.9528 253.093 57.7361 254.084 57.0551L254.704 58.2623C253.341 59.098 252.566 59.3766 251.389 59.3766C248.848 59.3766 247.113 57.4266 247.113 54.548C247.113 51.6385 248.848 49.7194 251.451 49.7194C253.589 49.7194 254.983 51.2361 254.983 53.5266V54.4242L248.693 54.3623ZM251.327 51.0504C250.057 51.0504 249.158 51.8242 248.817 53.248H253.31C253.279 51.8242 252.566 51.0504 251.327 51.0504Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M259.32 59.3456C258.267 59.3456 257.554 58.5099 257.554 57.3028V51.2051H256.005V50.3075L257.554 49.9361V47.2432H259.072V49.9051H261.489V51.1742H259.072V56.4051C259.072 56.498 259.103 57.2409 259.134 57.4266C259.165 57.6742 259.351 57.829 259.661 57.829C260.033 57.829 260.343 57.6742 260.993 57.179L261.613 58.2004C260.559 59.098 260.064 59.3456 259.32 59.3456Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M266.601 59.3456C266.229 59.3456 265.795 59.3147 265.331 59.2218C264.463 59.067 264.06 58.9123 263.162 58.448L263.782 57.179L264.246 57.4266C265.145 57.9218 266.012 58.0456 266.57 58.0456C267.499 58.0456 268.181 57.5504 268.181 56.8385C268.181 56.6218 268.088 56.3432 267.995 56.2194C267.623 55.6932 267.19 55.4147 265.579 54.7647C263.782 54.0218 263.193 53.4028 263.193 52.2885C263.193 50.679 264.525 49.6885 266.663 49.6885C267.035 49.6885 267.469 49.7194 267.84 49.7813C268.522 49.9051 268.863 50.029 269.637 50.4623L268.894 51.7004C268.305 51.329 268.057 51.2051 267.561 51.0813C267.19 50.9885 266.849 50.9266 266.508 50.9266C265.486 50.9266 264.804 51.3909 264.804 52.0718C264.804 52.2885 264.897 52.5051 265.021 52.629C265.393 53.0004 265.641 53.1242 267.283 53.8361C269.234 54.6409 269.761 55.2599 269.761 56.7147C269.823 58.3551 268.615 59.3456 266.601 59.3456Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M275.679 59.3456C275.307 59.3456 274.873 59.3147 274.408 59.2218C273.541 59.067 273.138 58.9123 272.24 58.448L272.859 57.179L273.324 57.4266C274.223 57.9218 275.09 58.0456 275.648 58.0456C276.577 58.0456 277.259 57.5504 277.259 56.8385C277.259 56.6218 277.166 56.3432 277.073 56.2194C276.701 55.6932 276.267 55.4147 274.656 54.7647C272.859 54.0218 272.271 53.4028 272.271 52.2885C272.271 50.679 273.603 49.6885 275.741 49.6885C276.113 49.6885 276.546 49.7194 276.918 49.7813C277.6 49.9051 277.94 50.029 278.715 50.4623L277.971 51.7004C277.383 51.329 277.135 51.2051 276.639 51.0813C276.267 50.9885 275.927 50.9266 275.586 50.9266C274.563 50.9266 273.882 51.3909 273.882 52.0718C273.882 52.2885 273.975 52.5051 274.099 52.629C274.47 53.0004 274.718 53.1242 276.36 53.8361C278.312 54.6409 278.839 55.2599 278.839 56.7147C278.87 58.3551 277.662 59.3456 275.679 59.3456Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M282.278 62.6575C281.906 63.3075 281.72 63.5551 281.132 64.267L280.109 63.4313C280.791 62.4718 280.977 62.1932 281.287 61.5742C281.627 60.8932 281.782 60.2123 281.782 59.2218V49.9361H283.3V60.6147C282.867 61.6051 282.681 62.0075 282.278 62.6575ZM282.526 47.6766C281.968 47.6766 281.534 47.2432 281.534 46.6861C281.534 46.129 281.968 45.6956 282.526 45.6956C283.084 45.6956 283.517 46.129 283.517 46.6861C283.517 47.2432 283.084 47.6766 282.526 47.6766Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M292.812 59.098V57.2099C291.728 58.6956 290.55 59.4075 289.156 59.4075C287.545 59.4075 286.708 58.5099 286.708 56.8385V49.967H288.227V56.467C288.227 57.6123 288.536 58.0766 289.311 58.0766C290.271 58.0766 291.728 57.0551 292.781 55.6004V49.998H294.33V59.1599H292.812V59.098Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M297.645 59.098V44.3337H299.163V59.098H297.645ZM303.005 59.098L299.225 54.1147L303.005 49.9361H304.988L301.084 54.1456L305.019 59.098H303.005Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M307.466 54.3623C307.466 56.5908 308.613 57.9528 310.472 57.9528C311.277 57.9528 311.866 57.7361 312.857 57.0551L313.477 58.2623C312.114 59.098 311.339 59.3766 310.162 59.3766C307.621 59.3766 305.886 57.4266 305.886 54.548C305.886 51.6385 307.621 49.7194 310.224 49.7194C312.362 49.7194 313.756 51.2361 313.756 53.5266V54.4242L307.466 54.3623ZM310.1 51.0504C308.83 51.0504 307.931 51.8242 307.59 53.248H312.083C312.052 51.8242 311.339 51.0504 310.1 51.0504Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M322.152 59.098V52.598C322.152 51.979 322.121 51.7932 321.997 51.5147C321.842 51.2051 321.501 51.0194 321.099 51.0194C320.138 51.0194 319.147 51.7004 317.567 53.4647V59.067H316.049V44.3028H317.567V51.8242C318.186 51.1123 318.372 50.9266 318.806 50.5551C319.519 49.9671 320.417 49.6575 321.285 49.6575C322.896 49.6575 323.701 50.4932 323.701 52.1956V59.067L322.152 59.098Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M333.089 59.098V57.2099C332.004 58.6956 330.827 59.4075 329.433 59.4075C327.822 59.4075 326.985 58.5099 326.985 56.8385V49.967H328.503V56.467C328.503 57.6123 328.813 58.0766 329.588 58.0766C330.548 58.0766 332.004 57.0551 333.058 55.6004V49.998H334.607V59.1599H333.089V59.098Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M340.927 59.3456C340.555 59.3456 340.122 59.3147 339.657 59.2218C338.789 59.067 338.387 58.9123 337.488 58.448L338.108 57.179L338.573 57.4266C339.471 57.9218 340.339 58.0456 340.896 58.0456C341.826 58.0456 342.507 57.5504 342.507 56.8385C342.507 56.6218 342.414 56.3432 342.321 56.2194C341.95 55.6932 341.516 55.4147 339.905 54.7647C338.108 54.0218 337.519 53.4028 337.519 52.2885C337.519 50.679 338.851 49.6885 340.989 49.6885C341.361 49.6885 341.795 49.7194 342.166 49.7813C342.848 49.9051 343.189 50.029 343.963 50.4623L343.22 51.7004C342.631 51.329 342.383 51.2051 341.919 51.0813C341.547 50.9885 341.206 50.9266 340.865 50.9266C339.843 50.9266 339.161 51.3909 339.161 52.0718C339.161 52.2885 339.254 52.5051 339.378 52.629C339.75 53.0004 339.998 53.1242 341.64 53.8361C343.592 54.6409 344.118 55.2599 344.118 56.7147C344.149 58.3551 342.941 59.3456 340.927 59.3456Z',
+      fill: 'themed',
+    },
+  ],
+};
