+++ hviktor-angular/projects/hviktor/src/logo/logos/logo-hbe.ts
@@ -1,0 +1,41 @@
+import { buildLogo } from '../logo-shared';
+import { LogoDefinition } from '../logo-types';
+
+export const LOGO_HBE: LogoDefinition = buildLogo(
+  'Helse Bergen logo',
+  [
+    // B
+    {
+      d: 'M207.347 43.9684H197.525V19.8636H206.251C209.391 19.8636 210.962 20.2654 212.276 21.4707C213.444 22.4934 214.175 24.21 214.175 25.7802C214.175 27.1316 213.663 28.556 212.787 29.5784C211.728 30.7836 210.743 31.2219 208.844 31.3315C212.495 31.4046 214.795 33.7786 214.795 37.5402C214.795 41.704 212.166 43.9684 207.347 43.9684ZM205.339 23.5157H202.345V29.7248H205.083C207.676 29.7248 209.063 28.6291 209.063 26.584C209.063 24.6483 207.676 23.516 205.339 23.516V23.5157ZM205.339 33.6325H202.345V40.2798H205.339C206.763 40.2798 207.42 40.1703 208.113 39.8416C209.136 39.3302 209.756 38.271 209.756 37.0292C209.756 34.8378 208.186 33.6325 205.339 33.6325Z',
+      fill: 'themed',
+    },
+    // E
+    {
+      d: 'M219.323 43.9684V19.8636H234.001V23.7714H223.924V29.7245H232.175V33.6322H223.924V40.0604H234.476V43.9681H219.323L219.323 43.9684Z',
+      fill: 'themed',
+    },
+    // R
+    {
+      d: 'M251.417 43.9684L245.904 34.3265H243.604V43.9684H239.003V19.8636H247.036C250.541 19.8636 252.148 20.3019 253.572 21.6533C254.886 22.8586 255.543 24.6117 255.543 26.8762C255.543 28.7389 255.105 30.2361 254.229 31.4049C253.316 32.5736 252.476 33.085 250.468 33.6328L256.456 43.9687H251.417V43.9684ZM246.598 23.7714H243.604V30.4187H246.598C249.519 30.4187 250.797 29.3961 250.797 26.9855C250.797 24.6845 249.592 23.7714 246.598 23.7714Z',
+      fill: 'themed',
+    },
+    // G
+    {
+      d: 'M277.78 43.6397C275.224 44.2241 273.727 44.4433 271.719 44.4433C263.723 44.4433 258.173 39.3299 258.173 31.8794C258.173 28.227 259.56 25.0497 262.153 22.6757C264.635 20.4115 267.41 19.4619 271.573 19.4619C273.325 19.4619 275.078 19.6445 276.611 20.0098C277.78 20.302 278.51 20.5576 280.189 21.2878L279.094 25.1228C278.218 24.7941 277.56 24.5749 277.122 24.4288C275.406 23.8445 273.69 23.5523 272.157 23.5523C266.753 23.5523 262.992 26.8759 262.992 31.6602C262.992 36.5907 267.118 40.3526 272.485 40.3526C273.581 40.3526 274.201 40.243 275.625 39.8413V33.1939H280.226V43.0551L277.78 43.6394V43.6397Z',
+      fill: 'themed',
+    },
+    // E
+    {
+      d: 'M286.178 43.9684V19.8636H300.856V23.7714H290.779V29.7245H299.03V33.6322H290.779V40.0604H301.331V43.9681H286.178L286.178 43.9684Z',
+      fill: 'themed',
+    },
+    // N
+    {
+      d: 'M321.923 44.1876L311.59 28.5925L310.787 27.1316C310.641 26.8394 310.568 26.6933 310.458 26.2915V43.9684H306.077V19.8636H311.371L321.119 35.13C321.813 36.2257 322.105 36.9927 322.397 37.7597V19.8636H326.743V44.1876H321.923Z',
+      fill: 'themed',
+    },
+  ],
+  '0 0 330 65',
+  330,
+  65,
+);
