+++ hviktor-angular/projects/hviktor/src/logo/logos/logo-hve.ts
@@ -1,0 +1,67 @@
+import { LogoDefinition } from '../logo-types';
+
+export const LOGO_HVE: LogoDefinition = {
+  viewBox: '0 0 240 61',
+  width: 240,
+  height: 61,
+  label: 'Helse Vest logo',
+  paths: [
+    // Accent dots (center)
+    {
+      d: 'M136.787 7.41402C136.787 11.5106 133.46 14.8302 129.357 14.8302C125.254 14.8302 121.932 11.5106 121.932 7.41402C121.932 3.31748 125.259 0 129.357 0C133.456 0 136.787 3.31964 136.787 7.41402Z',
+      fill: 'accent',
+    },
+    {
+      d: 'M136.787 30.0374C136.787 34.1318 133.46 37.4493 129.357 37.4493C125.254 37.4493 121.932 34.1297 121.932 30.0374C121.932 25.9452 125.259 22.6188 129.357 22.6188C133.456 22.6188 136.787 25.9409 136.787 30.0374Z',
+      fill: 'accent',
+    },
+    {
+      d: 'M159.48 30.0374C159.48 34.1318 156.153 37.4493 152.053 37.4493C147.952 37.4493 144.625 34.1297 144.625 30.0374C144.625 25.9452 147.95 22.6188 152.053 22.6188C156.156 22.6188 159.48 25.9409 159.48 30.0374Z',
+      fill: 'accent',
+    },
+    {
+      d: 'M136.787 52.702C136.787 56.7942 133.46 60.116 129.357 60.116C125.254 60.116 121.932 56.7942 121.932 52.702C121.932 48.6098 125.259 45.2833 129.357 45.2833C133.456 45.2833 136.787 48.6033 136.787 52.702Z',
+      fill: 'accent',
+    },
+    // Themed dot
+    {
+      d: 'M114.11 30.0374C114.11 34.1318 110.785 37.4493 106.682 37.4493C102.579 37.4493 99.2563 34.1297 99.2563 30.0374C99.2563 25.9452 102.581 22.6188 106.682 22.6188C110.783 22.6188 114.11 25.9409 114.11 30.0374Z',
+      fill: 'themed',
+    },
+    // HELSE
+    {
+      d: 'M14.3338 40.0273V30.8515H2.49643V40.0273H0V19.9711H2.49643V28.5712H14.3338V19.9711H16.8302V40.0273H14.3338Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M23.8173 40.0273V19.9711H35.1068V22.0675H26.3115V28.7529H33.5835V30.8493H26.3115V37.9287H35.3809V40.0251H23.8173V40.0273Z',
+      fill: 'themed',
+    },
+    { d: 'M40.5774 40.0273V19.9711H43.0738V37.9309H52.0511V40.0273H40.5774Z', fill: 'themed' },
+    {
+      d: 'M66.1307 23.374C65.0962 22.768 64.6688 22.5535 64.0311 22.3411C63.2707 22.0982 62.0542 21.9165 61.2606 21.9165C58.979 21.9165 57.4253 23.0413 57.4253 24.744C57.4253 26.2322 58.3085 26.9914 62.3872 29.1491C63.757 29.8777 64.8222 30.5475 65.5523 31.1864C66.5253 32.0069 67.1038 33.3747 67.1038 34.8344C67.1038 38.2658 64.5482 40.4257 60.5002 40.4257C59.1303 40.4257 57.8525 40.1827 56.7873 39.7562C56.0269 39.4519 55.5095 39.1498 54.2013 38.2658L55.4788 36.292C56.6053 37.0534 57.092 37.327 57.7933 37.6003C58.6746 37.9331 59.6499 38.1148 60.6843 38.1148C63.058 38.1148 64.4893 36.99 64.4893 35.0752C64.4893 34.4974 64.3972 34.0139 64.1845 33.6483C63.8513 33.0423 63.2707 32.4644 62.4814 32.0069C62.1463 31.8256 61.0197 31.2477 59.1346 30.2148C56.1211 28.604 54.8433 27.0242 54.8433 24.9585C54.8433 21.7679 57.491 19.6405 61.4776 19.6405C62.6348 19.6405 63.8206 19.7937 64.8572 20.1265C65.6794 20.4001 66.1657 20.6431 67.2923 21.3126L66.1351 23.3783L66.1307 23.374Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M72.6051 40.0273V19.9711H83.8947V22.0675H75.0994V28.7529H82.3713V30.8493H75.0994V37.9287H84.1687V40.0251H72.6051V40.0273Z',
+      fill: 'themed',
+    },
+    // VEST
+    {
+      d: 'M182.5 40.6114H179.819L172.365 19.9773H176.565L181.161 33.5298C181.253 33.8341 181.343 34.169 181.373 34.4729C181.373 34.1996 181.406 33.9851 181.558 33.5298L185.939 19.9773H189.956L182.502 40.6114H182.5Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M193.079 40.0338V19.9776H205.314V23.2295H196.915V28.1818H203.792V31.4337H196.915V36.7841H205.71V40.0338H193.079Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M220.886 24.7505C218.753 23.8073 217.019 23.3214 215.681 23.3214C213.885 23.3214 212.728 24.0503 212.728 25.1751C212.728 26.0263 213.217 26.5426 214.646 27.3022C217.719 28.9433 219.726 30.2191 220.671 31.1601C221.523 32.0113 222.041 33.3484 222.041 34.7468C222.041 38.3336 219.424 40.43 214.979 40.43C214.372 40.43 213.763 40.3994 213.184 40.3074C211.755 40.1257 211.085 39.8831 208.985 38.909L207.676 38.303L209.287 34.7775C210.749 35.5674 211.356 35.8716 212.15 36.1759C213.153 36.5415 214.249 36.7535 215.04 36.7535C216.805 36.7535 217.962 35.9657 217.962 34.749C217.962 33.8978 217.504 33.3221 216.196 32.5913L213.092 30.8276C209.531 28.7922 208.435 27.4554 208.435 25.2671C208.435 24.2645 208.739 23.2316 209.226 22.4724C209.559 21.9561 210.079 21.4395 210.718 20.9842C212.178 19.9819 213.425 19.6473 215.77 19.6473C217.109 19.6473 218.295 19.8006 219.544 20.1636C220.518 20.4679 221.066 20.6802 222.407 21.3169L220.886 24.7505Z',
+      fill: 'themed',
+    },
+    {
+      d: 'M233.69 23.2295V40.0338H229.854V23.2295H223.829V19.9776H239.715V23.2295H233.69Z',
+      fill: 'themed',
+    },
+  ],
+};
