+++ hviktor-angular/projects/hviktor/src/logo/logos/logo-hvikt.ts
@@ -1,0 +1,43 @@
+import { buildLogo } from '../logo-shared';
+import { LogoDefinition } from '../logo-types';
+
+export const LOGO_HVIKT: LogoDefinition = buildLogo(
+  'Helse Vest IKT logo',
+  [
+    // V
+    {
+      d: 'M203.757 44.9159H200.593L191.784 20.4796H196.746L202.175 36.5306C202.283 36.8905 202.391 37.2864 202.427 37.6463C202.427 37.3224 202.463 37.0704 202.643 36.5306L207.821 20.4796H212.567L203.757 44.9159Z',
+      fill: 'themed',
+    },
+    // E
+    {
+      d: 'M215.911 44.2324V20.4799H230.365V24.3306H220.442V30.1966H228.568V34.0473H220.442V40.3815H230.833V44.2321H215.911L215.911 44.2324Z',
+      fill: 'themed',
+    },
+    // S
+    {
+      d: 'M248.415 26.1301C245.898 25.0144 243.848 24.4386 242.266 24.4386C240.145 24.4386 238.779 25.3023 238.779 26.6339C238.779 27.6417 239.354 28.2535 241.044 29.1532C244.675 31.0967 247.049 32.6083 248.163 33.7237C249.17 34.7314 249.781 36.315 249.781 37.9705C249.781 42.217 246.689 44.7003 241.44 44.7003C240.72 44.7003 240.001 44.6643 239.318 44.5564C237.629 44.3404 236.837 44.0525 234.356 42.9008L232.81 42.181L234.716 38.0062C236.442 38.9419 237.161 39.3018 238.096 39.6617C239.282 40.0936 240.577 40.3455 241.512 40.3455C243.597 40.3455 244.963 39.4098 244.963 37.9702C244.963 36.9625 244.424 36.2787 242.878 35.4149L239.21 33.3275C235.003 30.9161 233.709 29.3326 233.709 26.7416C233.709 25.5539 234.068 24.3303 234.644 23.4308C235.039 22.819 235.651 22.2072 236.406 21.6673C238.132 20.4797 239.606 20.0838 242.374 20.0838C243.956 20.0838 245.359 20.2637 246.833 20.6956C247.983 21.0555 248.631 21.3074 250.212 22.0632L248.415 26.1301Z',
+      fill: 'themed',
+    },
+    // T
+    {
+      d: 'M263.194 24.3306V44.2324H258.663V24.3306H251.544V20.4799H270.313V24.3306H263.194Z',
+      fill: 'themed',
+    },
+    // I
+    { d: 'M283.545 44.2324V20.4799H286.493V44.2324H283.545Z', fill: 'themed' },
+    // K
+    {
+      d: 'M294.331 44.2324V20.4799H297.28V44.2324H294.331ZM306.988 44.2324L297.496 32.2481L306.772 20.4799H310.152L300.875 32.0324L310.619 44.2324H306.988Z',
+      fill: 'themed',
+    },
+    // T
+    {
+      d: 'M322.089 22.963V44.2324H319.141V22.963H311.231V20.4796H330V22.963H322.089Z',
+      fill: 'themed',
+    },
+  ],
+  '0 0 336 65',
+  336,
+  65,
+);
