+++ hviktor-angular/src/app/demo/pages/components/logo/logo-demo.ts
@@ -46,6 +46,10 @@ import { DemoPageComponent, DemoSectionComponent } from '../../../shared';
             <h3 hviHeading>hbe</h3>
             <hvi-logo company="hbe" [size]="selectedSize()" />
           </div>
+          <div>
+            <h3 hviHeading>hbe-hus</h3>
+            <hvi-logo company="hbe-hus" [size]="selectedSize()" />
+          </div>
         </div>
         <p class="mt-4 text-sm">Bytt fargetema i headeren for å se logoen i dark mode.</p>
       </app-demo-section>
